fix tox python3 overrides
We want to default to running all tox environments under python 3, so
set the basepython value in each environment.

We do not want to specify a minor version number, because we do not
want to have to update the file every time we upgrade python.

We do not want to set the override once in testenv, because that
breaks the more specific versions used in default environments like
py35 and py36.

Change-Id: I43c0875f0c9e3ea2e26fec844274b3d1ec769c8f
Signed-off-by: Doug Hellmann <doug@doughellmann.com>